BACKEND/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
python-dateutil
